+++ src/main/java/com/makersacademy/acebook/repository/CommentRepository.java
@@ -3,5 +3,8 @@ package com.makersacademy.acebook.repository;
 import com.makersacademy.acebook.model.Comment;
 import org.springframework.data.repository.CrudRepository;
 
+import java.util.List;
+
 public interface CommentRepository extends CrudRepository<Comment, Long> {
+    public List<Comment> findAllByPostId(Long postId);
 }
