+++ src/test/java/com/makersacademy/aceboook/controller/CommentControllerTest.java
@@ -56,8 +56,10 @@ public class CommentControllerTest {
         driver.findElement(By.id("submitComment")).click();
 
         WebElement commentsUL = driver.findElement(By.tagName("ul"));
+        System.out.println(commentsUL.getText());
         List<WebElement> commentsLI = commentsUL.findElements(By.tagName("li"));
-        String comment = commentsLI.get(0).getText();
+        System.out.println(commentsLI);
+        String comment = commentsLI.get(commentsLI.size() - 1).getText();
 
         Assert.assertEquals("Here is my new comment!", comment);
